Possible mismatch of charset/collation between columns on mysql/mariadb
Closes #45597
