+++ model/jpa/src/main/resources/META-INF/jpa-changelog-26.5.0.xml
@@ -166,7 +166,7 @@
         <dropIndex tableName="OFFLINE_USER_SESSION" indexName="IDX_OFFLINE_USS_BY_LAST_SESSION_REFRESH"/>
     </changeSet>
 
-    <changeSet author="keycloak" id="26.5.0-org-id-charset-mysql-fixed">
+    <changeSet author="keycloak" id="26.5.0-mysql-mariadb-default-charset-collation">
         <preConditions onFail="MARK_RAN" onSqlOutput="TEST">
             <and>
                 <or>
@@ -178,13 +178,14 @@
                 </not>
             </and>
         </preConditions>
-        <customChange class="org.keycloak.connections.jpa.updater.liquibase.custom.JpaUpdate26_5_0_OrgTableCharset"/>
+        <customChange class="org.keycloak.connections.jpa.updater.liquibase.custom.JpaUpdate26_5_0_OrgTablesSetDefaultCharsetAndCollation"/>
     </changeSet>
 
-    <changeSet author="keycloak" id="26.5.0-invitations-table-fixed">
+    <changeSet author="keycloak" id="26.5.0-invitations-table-fixed2">
         <preConditions onFail="MARK_RAN" onSqlOutput="TEST">
             <not>
-                <changeSetExecuted id="26.5.0-invitations-table" author="keycloak" changeLogFile="META-INF/jpa-changelog-26.5.0.xml"/>
+                <tableExists schemaName="${database.defaultSchemaName}"
+                             tableName="ORG_INVITATION"/>
             </not>
         </preConditions>
 
@@ -206,18 +207,6 @@
             <column name="EXPIRES_AT" type="INT"/>
             <column name="INVITE_LINK" type="VARCHAR(2048)"/>
         </createTable>
-
-        <modifySql dbms="mysql,mariadb">
-            <append value=" CHARACTER SET utf8mb3 COLLATE utf8mb3_unicode_ci"/>
-        </modifySql>
-    </changeSet>
-
-    <changeSet author="keycloak" id="26.5.0-invitations-table-fixed-fk">
-        <preConditions onFail="MARK_RAN" onSqlOutput="TEST">
-            <not>
-                <changeSetExecuted id="26.5.0-invitations-table" author="keycloak" changeLogFile="META-INF/jpa-changelog-26.5.0.xml"/>
-            </not>
-        </preConditions>
 
         <addForeignKeyConstraint baseTableName="ORG_INVITATION"
                                  baseColumnNames="ORGANIZATION_ID"
@@ -243,4 +232,5 @@
                              columnNames="ORGANIZATION_ID,EMAIL"
                              constraintName="UK_ORG_INVITATION_EMAIL"/>
     </changeSet>
+
 </databaseChangeLog>
