+++ model/jpa/src/main/resources/META-INF/jpa-changelog-26.6.0.xml
@@ -17,16 +17,6 @@
   -->
 <databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.1.xsd">
 
-    <changeSet author="keycloak" id="26.6.0-converge-org-charset">
-        <preConditions onFail="MARK_RAN">
-            <or>
-                <dbms type="mysql"/>
-                <dbms type="mariadb"/>
-            </or>
-        </preConditions>
-        <customChange class="org.keycloak.connections.jpa.updater.liquibase.custom.JpaUpdate26_6_0_MySQL_OrgTableCharset"/>
-    </changeSet>
-
     <changeSet author="keycloak" id="26.6.0-45009-broker-link-user-id">
         <preConditions onSqlOutput="TEST" onFail="MARK_RAN">
             <not>
